requirements.txt
python-telegram-bot==13.15
environs==11.0.0
google-cloud-dialogflow==2.31.0
google-cloud-api-keys==0.5.11